Tweaked compass zeroing, and changed camera matrix
Now zeroed (or any-number-ed) using Compass.heading = 0, rather than
`Compass.initializeZeroOffset`. Also, `Robot.compassHeading` was deprecated
and unused, so it was removed.

Camera matrix was not working before because the pyeuclid implementation
of euler angles is not consistent with my interpretation - it applies bank
around the x-axis, not the z-axis. For transparency, euler angles are no
longer used, and rotatex is used to account for the camera tilt (estimated
at 10 degrees - CHECK!). Height of camera is also now taken into account.