Added a class for the laser.
Henry should have ordered this, and will come in sometime during the week.
It'll need to be powered from 3.3V, which the mbed gives out. Either
plug it directly into 3.3V, or wire it up so that it can be turned on and
off. If you make it controllable, CHECK IF THE MBED CAN SUPPLY ENOUGH
CURRENT FIRST. If not, then use a FET.

Some kind of detector will be needed as well. An LDR, phototransistor, or
photodiode will work, as long as the mbed can tell if the laser is pointed
at it.