Tried to make it so that we can have separate commands for the robot to execute at different time during the match.
This is done in such a way that only one thread can be in control of the robot at once. At any point, a thread can pull the robot out from under the original thread, throwing an exception upon the next attribute lookup. This prevents two threads controlling the robot at the same time